Move helping functions to base.py
- These functions can be used in other tests as well and hence should be available to all.
