+++ tests/base.py
@@ -5,6 +5,7 @@ from unittest import mock
 
 from flask import g
 from flask_testing import TestCase
+from werkzeug.datastructures import Headers
 
 from database import create_session
 from mod_auth.models import Role, User
@@ -109,6 +110,39 @@ signup_information = {
     'existing_user_pwd': 'dummy_pwd',
     'existing_user_role': Role.user
 }
+
+
+def generate_signature(data, private_key):
+    """
+    Generate signature token of hook request
+
+    :param data: Signature's data
+    :param private_key: Signature's token
+    """
+    import hashlib
+    import hmac
+    algorithm = hashlib.__dict__.get('sha1')
+    encoded_key = bytes(private_key, 'latin-1')
+    mac = hmac.new(encoded_key, msg=data, digestmod=algorithm)
+    return mac.hexdigest()
+
+
+def generate_git_api_header(event, sig):
+    """
+    Create header for Github API Request, based on header information from https://developer.github.com/webhooks/.
+
+    :param event: Name of the event type that triggered the delivery.
+    :param sig: The HMAC hex digest of the response body. The HMAC hex digest is generated
+                using the sha1 hash function and the secret as the HMAC key.
+    """
+    return Headers([
+        ('X-GitHub-Event', event),
+        ('X-Github-Delivery', '72d3162e-cc78-11e3-81ab-4c9367dc0958'),
+        ('X-Hub-Signature', 'sha1={0}'.format(sig)),
+        ('User-Agent', 'GitHub-Hookshot/044aadd'),
+        ('Content-Type', 'application/json'),
+        ('Content-Length', 6615)
+    ])
 
 
 class BaseTestCase(TestCase):
